// src/api/orval/model/overviewTotalsResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { OverviewMetaDto } from "./overviewMetaDto";
import type { TotalsPayloadDto } from "./totalsPayloadDto";

export interface OverviewTotalsResponse {
  meta?: OverviewMetaDto;
  totals?: TotalsPayloadDto;
}